Move Block types out of `index.d.ts`
