+++ dotcom-rendering/src/components/LiveBlogRenderer.tsx
@@ -1,6 +1,7 @@
 import { Hide } from '@guardian/source/react-components';
 import type { ArticleFormat } from '../lib/articleFormat';
 import type { EditionId } from '../lib/edition';
+import type { Block } from '../types/blocks';
 import type { ServerSideTests, Switches } from '../types/config';
 import type { TagType } from '../types/tag';
 import { useConfig } from './ConfigContext';
