+++ dotcom-rendering/src/lib/getLiveblogAdPositions.ts
@@ -1,3 +1,4 @@
+import type { Block } from '../types/blocks';
 import {
 	calculateApproximateBlockHeight,
 	shouldDisplayAd,
