+++ dotcom-rendering/src/types/frontend.ts
@@ -1,6 +1,7 @@
 import type { SharedAdTargeting } from '../lib/ad-targeting';
 import type { EditionId } from '../lib/edition';
 import type { FEArticleBadgeType } from './badge';
+import type { Block } from './blocks';
 import type {
 	CommercialProperties,
 	ReaderRevenuePositions,
